requirements.txt
Django==3.2.5
dj-database-url==0.5.0
psycopg2-binary==2.9.1
djangorestframework==3.12.4
django-filter==2.4.0
drf-nested-routers==0.93.3
factory-boy==3.2.0
assertpy==1.1
